fix(lsp): no explicit bg on floating diagnostics
When a background color is set on diagnostic text highlight groups, and
the user changes the background of floating windows globally (ref.
sainnhe/sonokai#90), diagnostic text is displayed with a background
which doesn't match the one of the window.

Diagnostic text now inherits the background of the floating window,
which is Neovim's default.